* commit '9b0bb4712f42be688a7555c194db310648dbbb87':
  PT-8923 - Add possibility to export Product-Streams
  PT-8923 - Remove obsolet Shopware 5.3 checks
  PT-8923 - Fix code style; clean up code

+++ Bootstrap.php
@@ -8,6 +8,7 @@
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Configuration;
 use Doctrine\ORM\Tools\SchemaTool;
+use DoctrineExtensions\Query\Mysql\GroupConcat;
 use Shopware\Commands\SwagImportExport\ExportCommand;
 use Shopware\Commands\SwagImportExport\ImportCommand;
 use Shopware\Commands\SwagImportExport\ProfilesCommand;
@@ -28,6 +29,10 @@ use Shopware\Components\SwagImportExport\Service\ProfileService;
 use Shopware\Components\SwagImportExport\Service\UnderscoreToCamelCaseService;
 use Shopware\Components\SwagImportExport\UploadPathProvider;
 use Shopware\Components\SwagImportExport\Utils\FileHelper;
+use Shopware\CustomModels\ImportExport\Expression;
+use Shopware\CustomModels\ImportExport\Logger as LoggerModel;
+use Shopware\CustomModels\ImportExport\Profile;
+use Shopware\CustomModels\ImportExport\Session;
 use Shopware\Setup\SwagImportExport\Exception\MinVersionException;
 use Shopware\Setup\SwagImportExport\Install\DefaultProfileInstaller;
 use Shopware\Setup\SwagImportExport\Install\InstallerInterface;
@@ -95,7 +100,7 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
     /**
      * Returns the version of the plugin as a string
      *
-     * @throws Exception
+     * @throws RuntimeException
      *
      * @return string
      */
@@ -106,7 +111,7 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
         if ($info) {
             return $info['currentVersion'];
         }
-        throw new Exception('The plugin has an invalid version file.');
+        throw new RuntimeException('The plugin has an invalid version file.');
     }
 
     /**
@@ -198,8 +203,6 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
      * @param string $oldVersion
      *
      * @throws MinVersionException
-     * @throws Exception
-     * @throws Zend_Db_Adapter_Exception
      *
      * @return array
      */
@@ -329,7 +332,7 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
         $config = $this->em->getConfiguration();
         $classLoader = new \Doctrine\Common\ClassLoader('DoctrineExtensions', $this->Path() . 'Components/');
         $classLoader->register();
-        $config->addCustomStringFunction('GroupConcat', 'DoctrineExtensions\Query\Mysql\GroupConcat');
+        $config->addCustomStringFunction('GroupConcat', GroupConcat::class);
 
         $this->get('loader')->registerNamespace('Shopware\Components', $this->Path() . 'Components/');
         $this->get('loader')->registerNamespace('Shopware\Commands', $this->Path() . 'Commands/');
@@ -343,7 +346,7 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
     public function getDataFactory()
     {
         if ($this->dataFactory === null) {
-            $this->dataFactory = Enlight_Class::Instance('Shopware\Components\SwagImportExport\Factories\DataFactory');
+            $this->dataFactory = Enlight_Class::Instance(DataFactory::class);
         }
 
         return $this->dataFactory;
@@ -355,7 +358,7 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
     public function getProfileFactory()
     {
         if ($this->profileFactory === null) {
-            $this->profileFactory = Enlight_Class::Instance('Shopware\Components\SwagImportExport\Factories\ProfileFactory');
+            $this->profileFactory = Enlight_Class::Instance(ProfileFactory::class);
         }
 
         return $this->profileFactory;
@@ -367,7 +370,7 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
     public function getFileIOFactory()
     {
         if ($this->fileIOFactory === null) {
-            $this->fileIOFactory = Enlight_Class::Instance('Shopware\Components\SwagImportExport\Factories\FileIOFactory');
+            $this->fileIOFactory = Enlight_Class::Instance(FileIOFactory::class);
         }
 
         return $this->fileIOFactory;
@@ -379,7 +382,7 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
     public function getDataTransformerFactory()
     {
         if ($this->dataTransformerFactory === null) {
-            $this->dataTransformerFactory = Enlight_Class::Instance('Shopware\Components\SwagImportExport\Factories\DataTransformerFactory');
+            $this->dataTransformerFactory = Enlight_Class::Instance(DataTransformerFactory::class);
         }
 
         return $this->dataTransformerFactory;
@@ -460,6 +463,9 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
         );
     }
 
+    /**
+     * @return ProfileService
+     */
     public function registerProfileService()
     {
         return new ProfileService(
@@ -497,7 +503,7 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
     public function injectBackendAceEditor(Enlight_Event_EventArgs $args)
     {
         /** @var Shopware_Controllers_Backend_Index $controller */
-        $controller = $args->getSubject();
+        $controller = $args->get('subject');
         $request = $controller->Request();
         $response = $controller->Response();
         $view = $controller->View();
@@ -815,6 +821,10 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
         }
 
         $resource = $this->em->getRepository(\Shopware\Models\User\Resource::class)->find($resourceId);
+        if ($resource === null) {
+            return;
+        }
+
         foreach ($resource->getPrivileges() as $privilege) {
             $this->em->remove($privilege);
         }
@@ -840,10 +850,10 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
     private function getDoctrineModels()
     {
         return [
-            $this->em->getClassMetadata('Shopware\CustomModels\ImportExport\Session'),
-            $this->em->getClassMetadata('Shopware\CustomModels\ImportExport\Logger'),
-            $this->em->getClassMetadata('Shopware\CustomModels\ImportExport\Profile'),
-            $this->em->getClassMetadata('Shopware\CustomModels\ImportExport\Expression'),
+            $this->em->getClassMetadata(Session::class),
+            $this->em->getClassMetadata(LoggerModel::class),
+            $this->em->getClassMetadata(Profile::class),
+            $this->em->getClassMetadata(Expression::class),
         ];
     }
 
@@ -874,7 +884,7 @@ final class Shopware_Plugins_Backend_SwagImportExport_Bootstrap extends Shopware
         $profiles = $connection->fetchAll('SELECT COUNT(id) as count, name FROM s_import_export_profile GROUP BY name');
 
         foreach ($profiles as $profile) {
-            if ($profile['count'] == 1) {
+            if ((int) $profile['count'] === 1) {
                 continue;
             }
